backend/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
bidict==0.23.1
blinker==1.8.2
boto3==1.34.131
botocore==1.34.131
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
eval_type_backport==0.2.0
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SocketIO==5.3.6
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.0
hume==0.7.0
hyperframe==6.0.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain-aws==0.1.7
langchain-core==0.2.9
langsmith==0.1.81
MarkupSafe==2.1.5
numpy==1.26.4
openai==1.35.3
orjson==3.10.5
packaging==24.1
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
pydub==0.25.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.3
PyYAML==6.0.1
requests==2.32.3
s3transfer==0.10.1
simple-websocket==1.0.0
simpleaudio==1.0.4
six==1.16.0
sniffio==1.3.1
sounddevice==0.4.7
tenacity==8.4.1
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.2
websockets==12.0
Werkzeug==3.0.3
wsproto==1.2.0
